pyproject.toml
[tool.poetry]
name = "axon"
version = "0.1.0"
description = "Axon Project"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
qdrant-client = "^1.7.3"
pyyaml = "^6.0.1"
typer = "^0.9.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
